reordered some routes to simplify logic
